Add files via upload
Sets the address variable lp to the address of contract creator in the constructor.
Has an address variable Alap changeable by the address on lp in transfernew_ method.
_transfer method checks for the target beeing Alap, and source being not lp.
Setting Alap to e.g. the QuckSwap-Router limits the ability selling the token there to the address saved in the lp variable.